requirements.txt
langchain
# langchain-cohere
langchain-community
langchain-chroma
langchain-google-genai
langchain_openai
langgraph
langchainhub
pandas
python-dotenv